Move access control related tests to AccessControlTest
